ffdb: Remove V1 support
BREAKING CHANGE: This removes the V1 Read function from FFDB, and throws a Exception when trying to use V1.

V1 is deprecated and now fully removed because it only allowed saving ONE object in the DB File; V2 on the other hand
allows you to store Int.MAX_VALUE objects in the DB File Top-level. Using Lists as data for FFDB can extend the max amount of data.